+++ package.json
@@ -9,7 +9,6 @@
   },
   "dependencies": {
     "firebase": "^6.0.2",
-    "firebase-admin": "^7.3.0",
     "material-design-icons": "^3.0.1",
     "react": "16.8.3",
     "react-loop": "^0.1.0",
