+++ screens/Home.js
@@ -19,7 +19,6 @@ class ViroSample extends Component {
     this.handleStart = this.handleStart.bind(this);
   }
 
-
   handleStart() {
     const { history, startGame } = this.props;
     startGame();
@@ -71,12 +70,19 @@ class ViroSample extends Component {
           >
             <Text style={localStyles.buttonText}>Instructions</Text>
           </TouchableHighlight>
-          <TouchableHighlight
+          {/* <TouchableHighlight
             style={localStyles.buttons}
             onPress={() => history.push('/login')}
             underlayColor="#04152b"
           >
             <Text style={localStyles.buttonText}>Login</Text>
+          </TouchableHighlight> */}
+          <TouchableHighlight
+            style={localStyles.buttons}
+            onPress={() => history.push('/signup')}
+            underlayColor="#04152b"
+          >
+            <Text style={localStyles.buttonText}>Signup</Text>
           </TouchableHighlight>
         </View>
       </View>
@@ -149,6 +155,6 @@ let localStyles = StyleSheet.create({
     shadowOffset: { width: 2, height: 3 },
     shadowOpacity: 0.4,
     shadowRadius: 3,
-    marginBottom: 15,
+    // marginBottom: 15,
   },
 });
